+++ web/src/lib/components/asset-viewer/detail-panel-tags.svelte
@@ -1,15 +1,13 @@
 <script lang="ts">
   import Badge from '$lib/components/elements/badge.svelte';
-  import Button from '$lib/components/elements/buttons/button.svelte';
   import Icon from '$lib/components/elements/icon.svelte';
-  import Combobox from '$lib/components/shared-components/combobox.svelte';
-  import FullScreenModal from '$lib/components/shared-components/full-screen-modal.svelte';
   import { AppRoute } from '$lib/constants';
   import { isSharedLink } from '$lib/utils';
-  import { getAllTags, getAssetInfo, type AssetResponseDto, type TagResponseDto } from '@immich/sdk';
-  import { mdiCameraIris, mdiPencil, mdiPlus, mdiTag, mdiTagOutline } from '@mdi/js';
+  import { getAssetInfo, type AssetResponseDto } from '@immich/sdk';
+  import { mdiPlus } from '@mdi/js';
   import { t } from 'svelte-i18n';
   import { tagAssets } from '$lib/utils/asset-utils';
+  import TagAssetForm from '$lib/components/forms/tag-asset-form.svelte';
 
   export let asset: AssetResponseDto;
   export let isOwner: boolean;
@@ -17,24 +15,20 @@
   $: tags = asset.tags || [];
 
   let isOpen = false;
-  let selectedTags = new Set<string>();
-  let availableTags: TagResponseDto[] = [];
 
   const handleAdd = async () => {
-    availableTags = await getAllTags();
     isOpen = true;
   };
 
   const handleCancel = () => (isOpen = false);
 
-  const handleTag = async () => {
-    const ids = await tagAssets({ tagIds: [...selectedTags], assetIds: [asset.id] });
+  const handleTag = async (tagIds: string[]) => {
+    const ids = await tagAssets({ tagIds, assetIds: [asset.id], showNotification: false });
 
     if (ids) {
       isOpen = false;
     }
 
-    selectedTags.clear();
     asset = await getAssetInfo({ id: asset.id });
   };
 </script>
@@ -62,20 +56,5 @@
 {/if}
 
 {#if isOpen}
-  <FullScreenModal title={$t('tag_assets')} icon={mdiTag} onClose={handleCancel}>
-    <form on:submit|preventDefault={handleTag} autocomplete="off" id="create-tag-form">
-      <div class="my-4 flex flex-col gap-2">
-        <Combobox
-          on:select={({ detail: option }) => option && selectedTags.add(option.value)}
-          label={$t('tag')}
-          options={availableTags.map((tag) => ({ id: tag.id, label: tag.value, value: tag.id }))}
-          placeholder={$t('search_tags')}
-        />
-      </div>
-    </form>
-    <svelte:fragment slot="sticky-bottom">
-      <Button color="gray" fullwidth on:click={() => handleCancel()}>{$t('cancel')}</Button>
-      <Button type="submit" fullwidth form="create-tag-form">{$t('tag_assets')}</Button>
-    </svelte:fragment>
-  </FullScreenModal>
+  <TagAssetForm onTag={(tagsIds) => handleTag(tagsIds)} onCancel={handleCancel} />
 {/if}
